+++ darwin/exporter/formats/__init__.py
@@ -6,7 +6,7 @@ supported_formats: List[str] = [
     "coco",
     "cvat",
     "dataloop",
-    "darwin",
+    "darwin_1.0",
     "instance_mask",
     "pascalvoc",
     "semantic_mask",
