+++ darwin/version/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.7.21"
+__version__ = "0.7.22"
